+++ sampratistaana-javafx/src/main/java/org/sampratistaana/controllers/DonationListController.java
@@ -56,7 +56,9 @@ public class DonationListController extends BaseController{
 	}
 	
 	public void deleteDonation() throws IOException {
-		new CreditManager().deleteDonation(donationList.getSelectionModel().getSelectedItem());
-		loadForm("DonationList");
+		if(showDeleteConfirmation()) {
+			new CreditManager().deleteDonation(donationList.getSelectionModel().getSelectedItem());
+			loadForm("DonationList");
+		}
 	}
 }
